Refactored Grid
removed data structure where coordinate was a physical place in memory. Instead memory is a simple array of GridItem pointers which are accessed with the various maps in the static class of Grid. The GetCell and GetCells function now accept a lambda function of type ItemFunc in order for the operations to allow for filtering of certain types. This way items can be iterated over a single time instead of multiple.